Changes to mccore/mcdriver for gui app.
- ArrayND.copyTo() for efficient copying
- replace uint with size_t or unsigned int
- add shared atomic mccore::ion_counter. All threads advance this when starting a new ion, new value becomes ion_id
- add shared atomic mccore::abort_flag. Can be used to cancel a running simulation
- add shared mccore::tally_mutex to sync access to the tally by worker threads and outside threads
- add atomic thread_ion_counter, per thread count of simulated ion histories
- add tally::clone, tally::copyTo for efficient access by calling processes

+++ src/event_stream.cpp
@@ -84,16 +84,16 @@ int event_stream::merge(const event_stream& ev)
     fs_.open(fname_, std::ios::out | std::ios::binary | std::ios::app );
 
     // mem buffer ~1MB
-    uint n = (1 << 10);
+    size_t n = (1 << 10);
     std::vector<float> buff(n*cols());
 
     // copy data in chunks
-    uint nrows = ev.rows(); // total rows to copy
+    size_t nrows = ev.rows(); // total rows to copy
     while (nrows) {
-        uint n1 = std::min(nrows,n); // # of rows to copy in this iter
+        size_t n1 = std::min(nrows,n); // # of rows to copy in this iter
         nrows -= n1;
         rows_ += n1;
-        uint nbytes = n1*cols()*sizeof(float);
+        size_t nbytes = n1*cols()*sizeof(float);
 
         // read from raw buffer
         ifs.read((char*)buff.data(), nbytes);
